added contextmenuprovider.py
